Refactor webmention sending/getting logic
Alright, this is the critical one. For unit testing we really don't want
to send or receive web mentions to/from a real source, so we want to be
able to mock things out. This refactoring moves all the logic for
sending or receiving webmentions into a separate service class.
